Separate hlo_op_profiler into platform-specific files
