+++ xla/service/gpu/model/BUILD
@@ -1,7 +1,8 @@
 load("@bazel_skylib//rules:build_test.bzl", "build_test")
-load("@local_config_cuda//cuda:build_defs.bzl", "if_cuda")
 load("@local_config_rocm//rocm:build_defs.bzl", "if_rocm_is_configured")
 load("@rules_cc//cc:cc_library.bzl", "cc_library")
+load("//xla/stream_executor:build_defs.bzl", "if_cuda_or_rocm_is_configured")
+load("//xla/tsl/platform/default:cuda_build_defs.bzl", "if_cuda_is_configured")
 load("//xla:xla.default.bzl", "xla_cc_test")
 
 # Libraries for performance modeling of HLO.
@@ -745,11 +746,112 @@ xla_cc_test(
 )
 
 cc_library(
-    name = "hlo_op_profiler_lib",
-    srcs = ["hlo_op_profiler.cc"],
+    name = "hlo_op_profiler_cuda",
+    srcs = [
+        "hlo_op_profiler.cc",
+        "hlo_op_profiler_cuda.cc",
+    ],
+    hdrs = ["hlo_op_profiler.h"],
+    compatible_with = [],
+    tags = [
+        "gpu",
+        "manual",
+    ],
+    deps = [
+        ":hlo_op_profile_proto_cc",
+        "//xla:debug_options_flags",
+        "//xla:literal",
+        "//xla:shape_util",
+        "//xla:util",
+        "//xla:xla_data_proto_cc",
+        "//xla/hlo/ir:hlo",
+        "//xla/service:executable",
+        "//xla/service:gpu_plugin",
+        "//xla/service:hlo_module_config",
+        "//xla/service:hlo_runner_interface",
+        "//xla/service:hlo_verifier",
+        "//xla/service:interpreter_plugin",
+        "//xla/stream_executor:device_description",
+        "//xla/tests:test_utils",
+        "//xla/tsl/platform:errors",
+        "//xla/tsl/platform:statusor",
+        "@com_google_absl//absl/algorithm:container",
+        "@com_google_absl//absl/base:no_destructor",
+        "@com_google_absl//absl/base:nullability",
+        "@com_google_absl//absl/container:flat_hash_set",
+        "@com_google_absl//absl/log",
+        "@com_google_absl//absl/log:check",
+        "@com_google_absl//absl/log:die_if_null",
+        "@com_google_absl//absl/status:statusor",
+        "@com_google_absl//absl/time",
+        "@com_google_absl//absl/types:span",
+        "@tsl//tsl/platform:errors",
+        "@tsl//tsl/platform:statusor",
+        # keep sorted
+        "//xla/backends/profiler/gpu:cupti_buffer_events",
+        "//xla/backends/profiler/gpu:cupti_collector",
+        "//xla/backends/profiler/gpu:cupti_tracer",
+    ],
+)
+
+cc_library(
+    name = "hlo_op_profiler_rocm",
+    srcs = [
+        "hlo_op_profiler.cc",
+        "hlo_op_profiler_rocm.cc",
+    ],
+    hdrs = ["hlo_op_profiler.h"],
+    compatible_with = [],
+    tags = [
+        "gpu",
+        "manual",
+        "rocm-only",
+    ],
+    deps = [
+        ":hlo_op_profile_proto_cc",
+        "//xla:debug_options_flags",
+        "//xla:literal",
+        "//xla:shape_util",
+        "//xla:util",
+        "//xla:xla_data_proto_cc",
+        "//xla/hlo/ir:hlo",
+        "//xla/service:executable",
+        "//xla/service:gpu_plugin",
+        "//xla/service:hlo_module_config",
+        "//xla/service:hlo_runner_interface",
+        "//xla/service:hlo_verifier",
+        "//xla/service:interpreter_plugin",
+        "//xla/stream_executor:device_description",
+        "//xla/tests:test_utils",
+        "//xla/tsl/platform:errors",
+        "//xla/tsl/platform:statusor",
+        "@com_google_absl//absl/base:no_destructor",
+        "@com_google_absl//absl/base:nullability",
+        "@com_google_absl//absl/container:flat_hash_set",
+        "@com_google_absl//absl/log",
+        "@com_google_absl//absl/log:check",
+        "@com_google_absl//absl/log:die_if_null",
+        "@com_google_absl//absl/status:statusor",
+        "@com_google_absl//absl/strings",
+        "@com_google_absl//absl/time",
+        "@com_google_absl//absl/types:span",
+        "@tsl//tsl/platform:errors",
+        "@tsl//tsl/platform:statusor",
+        # keep sorted
+        "//xla/backends/profiler/gpu:rocm_collector",
+        "//xla/backends/profiler/gpu:rocm_tracer",
+    ],
+)
+
+cc_library(
+    name = "hlo_op_profiler_stub",
+    srcs = [
+        "hlo_op_profiler.cc",
+        "hlo_op_profiler_stub.cc",
+    ],
     hdrs = ["hlo_op_profiler.h"],
     compatible_with = get_compatible_with_portable(),
-    local_defines = if_cuda(["GOOGLE_CUDA"]) + if_rocm_is_configured(["TENSORFLOW_USE_ROCM"]),
+    tags = ["manual"],
     deps = [
         ":hlo_op_profile_proto_cc",
         "//xla:debug_options_flags",
@@ -779,15 +881,33 @@ cc_library(
         "@com_google_absl//absl/types:span",
         "@tsl//tsl/platform:errors",
         "@tsl//tsl/platform:statusor",
-    ] + if_cuda([
-        # keep sorted
-        "//xla/backends/profiler/gpu:cupti_buffer_events",
-        "//xla/backends/profiler/gpu:cupti_collector",
-        "//xla/backends/profiler/gpu:cupti_tracer",
-    ]) + if_rocm_is_configured([
-        "//xla/backends/profiler/gpu:rocm_collector",
-        "//xla/backends/profiler/gpu:rocm_tracer",
-    ]),
+    ],
+)
+
+cc_library(
+    name = "hlo_op_profiler_lib",
+    srcs = [],
+    hdrs = ["hlo_op_profiler.h"],
+    compatible_with = get_compatible_with_portable(),
+    deps = [
+        ":hlo_op_profile_proto_cc",
+        "//xla:xla_data_proto_cc",
+        "//xla/hlo/ir:hlo",
+        "//xla/service:hlo_runner_interface",
+        "//xla/stream_executor:device_description",
+        "@com_google_absl//absl/base:nullability",
+        "@com_google_absl//absl/container:flat_hash_set",
+        "@com_google_absl//absl/status:statusor",
+        "@com_google_absl//absl/time",
+        "@com_google_absl//absl/types:span",
+    ] + if_cuda_or_rocm_is_configured(
+        [],
+        [":hlo_op_profiler_stub"],
+    ) + if_rocm_is_configured(
+        [":hlo_op_profiler_rocm"],
+    ) + if_cuda_is_configured(
+        [":hlo_op_profiler_cuda"],
+    ),
 )
 
 xla_test(
@@ -849,7 +969,6 @@ xla_test(
     # TODO(b/372714955): Fix the memory leak!
     args = if_google(["--heap_check="]),
     backends = ["gpu"],
-    local_defines = if_cuda(["GOOGLE_CUDA"]) + if_rocm_is_configured(["TENSORFLOW_USE_ROCM"]),
     deps = [
         ":hlo_op_profiler_lib",
         "//xla:xla_data_proto_cc",
