+++ xla/service/gpu/model/hlo_op_profiler.cc
@@ -15,7 +15,6 @@ limitations under the License.
 
 #include "xla/service/gpu/model/hlo_op_profiler.h"
 
-#include <algorithm>
 #include <array>
 #include <cstdint>
 #include <memory>
@@ -31,7 +30,6 @@ limitations under the License.
 #include "absl/log/die_if_null.h"
 #include "absl/log/log.h"
 #include "absl/status/statusor.h"
-#include "absl/strings/match.h"
 #include "absl/time/clock.h"
 #include "absl/time/time.h"
 #include "absl/types/span.h"
@@ -52,15 +50,6 @@ limitations under the License.
 #include "xla/tsl/platform/statusor.h"
 #include "xla/util.h"
 #include "xla/xla_data.pb.h"
-
-#if defined(GOOGLE_CUDA)
-#include "xla/backends/profiler/gpu/cupti_buffer_events.h"
-#include "xla/backends/profiler/gpu/cupti_collector.h"
-#include "xla/backends/profiler/gpu/cupti_tracer.h"
-#elif defined(TENSORFLOW_USE_ROCM)
-#include "xla/backends/profiler/gpu/rocm_collector.h"
-#include "xla/backends/profiler/gpu/rocm_tracer.h"
-#endif
 
 namespace xla {
 namespace gpu {
@@ -229,154 +218,6 @@ const absl::flat_hash_set<HloOpcode>& HloOpProfiler::TooFastToMeasure() {
   return TooFastToMeasureOps();
 }
 
-#ifdef GOOGLE_CUDA
-class CuptiKernelTracer : public HloOpProfiler::KernelTracer,
-                          public profiler::CuptiTraceCollector {
- public:
-  CuptiKernelTracer()
-      : profiler::CuptiTraceCollector({}),
-        cupti_tracer_(profiler::CuptiTracer::GetCuptiTracerSingleton()) {
-    CHECK(cupti_tracer_->IsAvailable());
-    profiler::CuptiTracerOptions options;
-    options.cbids_selected.push_back(
-        // Not interested in API callbacks, but empty list enables them all.
-        CUPTI_DRIVER_TRACE_CBID_cu64GLMapBufferObject);
-    options.activities_selected.push_back(CUPTI_ACTIVITY_KIND_KERNEL);
-    cupti_tracer_->Enable(options, this).IgnoreError();
-  }
-
-  uint64_t getMedianKernelTimeNs() && override {
-    cupti_tracer_->Disable();  // Also flushes buffer.
-    if (kernel_times_ns_.empty()) {
-      LOG(ERROR) << "No kernel events";
-      return 0;
-    }
-    absl::c_sort(kernel_times_ns_);
-    auto i = kernel_times_ns_.size() / 2;
-    // Return median value if number of values is odd.
-    if (kernel_times_ns_.size() % 2 != 0) {
-      return kernel_times_ns_[i];
-    }
-    // Return average of the two middle values if the number of values is even.
-    return (kernel_times_ns_[i - 1] + kernel_times_ns_[i] + 1) / 2;
-  }
-
- private:
-  // CuptiTraceCollector
-  void AddEvent(profiler::CuptiTracerEvent&& event) override {
-    if (event.type == profiler::CuptiTracerEventType::Kernel) {
-      kernel_times_ns_.push_back(event.end_time_ns - event.start_time_ns);
-    }
-    VLOG(5) << "CuptiTracerEvent: " << event.name << ", "
-            << event.end_time_ns - event.start_time_ns << "ns";
-  }
-  void OnEventsDropped(const std::string& reason,
-                       uint32_t num_events) override {
-    LOG(WARNING) << "Dropped " << num_events << " events: " << reason;
-  }
-  void Flush() override {}
-
-  profiler::CuptiTracer* cupti_tracer_;
-  std::vector<uint64_t> kernel_times_ns_;
-};
-#endif  // GOOGLE_CUDA
-
-#ifdef TENSORFLOW_USE_ROCM
-class RocmKernelTracer : public HloOpProfiler::KernelTracer,
-                         public profiler::RocmTraceCollector {
- public:
-  RocmKernelTracer()
-      :  rocm_tracer_(&profiler::RocmTracer::GetRocmTracerSingleton()),
-         profiler::RocmTraceCollector(MakeCollectorOptions(
-             profiler::RocmTracer::GetRocmTracerSingleton().NumGpus())),
-         start_timestamp_ns_(profiler::RocmTracer::GetTimestamp()) {
-    CHECK(rocm_tracer_->IsAvailable());
-    profiler::RocmTracerOptions options;
-    options.max_annotation_strings = 1024 * 1024;
-    rocm_tracer_->Enable(options, this);
-  }
-
-  uint64_t getMedianKernelTimeNs() && override {
-    rocm_tracer_->Disable();  // Also flushes buffer.
-    if (kernel_times_ns_.empty()) {
-      LOG(ERROR) << "No kernel events";
-      return 0;
-    }
-    std::sort(kernel_times_ns_.begin(), kernel_times_ns_.end());
-    auto i = kernel_times_ns_.size() / 2;
-    // Return median value if number of values is odd.
-    if (kernel_times_ns_.size() % 2 != 0) {
-      return kernel_times_ns_[i];
-    }
-    // Return average of the two middle values if the number of values is even.
-    return (kernel_times_ns_[i - 1] + kernel_times_ns_[i] + 1) / 2;
-  }
-
- private:
-  static profiler::RocmTraceCollectorOptions MakeCollectorOptions(
-      uint32_t num_gpus) {
-    profiler::RocmTraceCollectorOptions options;
-    options.max_callback_api_events = 2 * 1024 * 1024;
-    options.max_activity_api_events = 2 * 1024 * 1024;
-    options.max_annotation_strings = 1024 * 1024;
-    options.num_gpus = num_gpus;
-    return options;
-  }
-
-  static bool IsComputeKernel(const std::string& name) {
-    // Exclude ROCm runtime (memory fill, copy, etc.)
-    return !absl::StartsWith(name, "__amd_rocclr_");
-  }
-
-  // RocmTraceCollector interface
-  void AddEvent(profiler::RocmTracerEvent&& event, bool is_auxiliary) override {
-    if (event.type == profiler::RocmTracerEventType::Kernel &&
-        event.source == profiler::RocmTracerEventSource::Activity &&
-        IsComputeKernel(event.name)) {
-      // Filter out events that started before the tracer was created.
-      // This discards stray events from warmup runs.
-      if (event.start_time_ns < start_timestamp_ns_) {
-        return;
-      }
-      kernel_times_ns_.push_back(event.end_time_ns - event.start_time_ns);
-      VLOG(3) << "Kernel dispatch: " << event.name << ", "
-              << event.end_time_ns - event.start_time_ns << "ns";
-    }
-  }
-
-  void OnEventsDropped(const std::string& reason,
-                       uint32_t num_events) override {
-    LOG(WARNING) << "Dropped " << num_events << " events: " << reason;
-  }
-  void Flush() override {}
-  void Export(tsl::profiler::XSpace* space) override {}
-
-  profiler::RocmTracer* rocm_tracer_;
-  std::vector<uint64_t> kernel_times_ns_;
-  uint64_t start_timestamp_ns_;
-};
-#endif  // TENSORFLOW_USE_ROCM
-
-#if !defined(GOOGLE_CUDA) && !defined(TENSORFLOW_USE_ROCM)
-class StubKernelTracer : public HloOpProfiler::KernelTracer {
- public:
-  uint64_t getMedianKernelTimeNs() && override {
-    LOG(FATAL) << "Not built with --config=cuda or --config=rocm";
-  }
-};
-#endif
-
-/*static*/ std::unique_ptr<HloOpProfiler::KernelTracer>
-HloOpProfiler::GetKernelTracer() {
-#if defined(GOOGLE_CUDA)
-  return std::make_unique<CuptiKernelTracer>();
-#elif defined(TENSORFLOW_USE_ROCM)
-  return std::make_unique<RocmKernelTracer>();
-#else
-  return std::make_unique<StubKernelTracer>();
-#endif
-}
-
 /*static*/ std::unique_ptr<HloModule> HloOpProfiler::MakeModuleForMeasurements(
     HloOpcode op, PrimitiveType data_type, int chain_length) {
   constexpr int64_t kInputSize = 1;
@@ -420,10 +261,6 @@ HloOpProfiler::GetKernelTracer() {
 
 absl::StatusOr<absl::Duration> HloOpProfiler::MeasureOpChainDuration(
     HloOpcode op, PrimitiveType data_type, int chain_length) {
-#if !defined(GOOGLE_CUDA) && !defined(TENSORFLOW_USE_ROCM)
-  return FailedPrecondition("Not built with --config=cuda or --config=rocm");
-#else
-
   std::unique_ptr<HloModule> module =
       MakeModuleForMeasurements(op, data_type, chain_length);
   HloVerifier verifier(/*layout_sensitive=*/true,
@@ -451,11 +288,10 @@ absl::StatusOr<absl::Duration> HloOpProfiler::MeasureOpChainDuration(
   TF_RETURN_IF_ERROR(
       runner_.ExecuteWithExecutable(ex.get(), args_small).status());
 
-#if defined(GOOGLE_CUDA)
-  CuptiKernelTracer kernel_tracer;
-#elif defined(TENSORFLOW_USE_ROCM)
-  RocmKernelTracer kernel_tracer;
-#endif
+  std::unique_ptr<KernelTracer> kernel_tracer = GetKernelTracer();
+  if (!kernel_tracer) {
+    return FailedPrecondition("Not built with --config=cuda or --config=rocm");
+  }
   for (int i = 0; i < 10; ++i) {  // Run a few times to reduce noise.
     TF_RETURN_IF_ERROR(
         runner_.ExecuteWithExecutable(ex.get(), args_small).status());
@@ -463,8 +299,7 @@ absl::StatusOr<absl::Duration> HloOpProfiler::MeasureOpChainDuration(
         runner_.ExecuteWithExecutable(ex.get(), args_large).status());
   }
 
-  return absl::Nanoseconds(std::move(kernel_tracer).getMedianKernelTimeNs());
-#endif  // !defined(GOOGLE_CUDA) && !defined(TENSORFLOW_USE_ROCM)
+  return absl::Nanoseconds(std::move(*kernel_tracer).getMedianKernelTimeNs());
 }
 
 HloOpProfiler::HloOpProfiler(HloRunnerInterface* const absl_nonnull runner,
